Add OTA (Over The Air) updating firmware from SD Card.
Add feature to update the module ESP32 Cam without serial USB Cable or FTDI (Future Technology Devices International).
The NodeMCU ESP32 load the new firmware if a new one exists in the path /firmware/mailbox.bin from the SD Card.
After the updating, the .bin is deleted and the ESP32 reboot and bootstrap with the new firmware.